+++ renderer/src/contexts/TabsContext.jsx
@@ -4,4 +4,5 @@ export const TabsContext = React.createContext({
   tabs: [{ id: "", name: "", messages: [""] }],
   addTab: (tab) => {},
   removeTab: (name) => {},
+  removeTabById: (id) => {},
 });