+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.10] - 2020-12-23
+
+- Added support for the new T-Chatter Browser Sync extension.
+- Fixed issue with message limit setting.
+- Added reset options button.
+
 ## [0.3.9] - 2020-12-21
 
 - Fixed tooltip not showing when hovering over an emote.
@@ -74,6 +80,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Release v0.1.0
 
 [unreleased]: https://github.com/T-Chatter/T-Chatter/compare/v0.3.9...dev
+[0.3.10]: https://github.com/T-Chatter/T-Chatter/compare/v0.3.9...v0.3.10
 [0.3.9]: https://github.com/T-Chatter/T-Chatter/compare/v0.3.8...v0.3.9
 [0.3.8]: https://github.com/T-Chatter/T-Chatter/compare/v0.3.7...v0.3.8
 [0.3.7]: https://github.com/T-Chatter/T-Chatter/compare/v0.3.6...v0.3.7